refactor: [EUDIW-202] Move CIE components to example app (#166)
