+++ package.json
@@ -56,7 +56,6 @@
     "@pagopa/eslint-config": "^3.0.0",
     "@pagopa/io-react-native-crypto": "^0.2.3",
     "@pagopa/io-react-native-jwt": "^1.3.0",
-    "@pagopa/react-native-cie": "^1.3.0",
     "@react-native/eslint-config": "^0.72.2",
     "@rushstack/eslint-patch": "^1.3.2",
     "@types/jest": "^28.1.2",
@@ -70,7 +69,6 @@
     "react": "18.2.0",
     "react-native": "0.72.14",
     "react-native-builder-bob": "^0.20.0",
-    "react-native-webview": "^13.10.5",
     "typed-openapi": "^0.4.1",
     "typescript": "^5.0.2"
   },
@@ -80,10 +78,8 @@
   "peerDependencies": {
     "@pagopa/io-react-native-crypto": "*",
     "@pagopa/io-react-native-jwt": "*",
-    "@pagopa/react-native-cie": "*",
     "react": "*",
-    "react-native": "*",
-    "react-native-webview": "*"
+    "react-native": "*"
   },
   "engines": {
     "node": ">= 16.0.0"
